move copied bluetooth_event.rs into mijia
