+++ bluez-generated/src/lib.rs
@@ -1,2 +1,1 @@
-pub mod bluetooth_event;
 pub mod generated;
